isolate user preference form validation
