+++ utils/enum/user-gender-activity.enum.ts
@@ -1,0 +1,11 @@
+export enum GenderEnum {
+    MALE = 'MALE',
+    FEMALE = 'FEMALE'
+}
+
+export enum PhysicalActivityEnum {
+    LOW = 'LOW',
+    SEDENTARY = 'SEDENTARY',
+    MODERATE = 'MODERATE',
+    HIGH = 'HIGH',
+}
